% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/acquire.R
\name{acquire}
\alias{acquire}
\title{Acquire player tracking data from http://stats.nba.com/}
\usage{
acquire(type = "shot", PlayerID = "2544", DateFrom = "", DateTo = "",
  GameSegment = "", LastNGames = "0", LeagueID = "00", Location = "",
  Month = "0", OpponentTeamID = "0", Outcome = "", Period = "0",
  Season = "2013-14", SeasonSegment = "", SeasonType = "Regular+Season",
  TeamID = "0", VsConference = "", VsDivision = "", PerMode = "PerGame")
}
\arguments{
\item{type}{Required type of tracking data. Eligible values are 'shot', 'rebound',
'pass', 'shotdefend', 'reb', 'shots'. The most useful/interesting are 'shot' and 'rebound'.}

\item{PlayerID}{Required player ID. See \code{data(players)}}

\item{DateFrom}{Optionally set date to begin acquiring data (in the format 'YYYY/MM/DD')}

\item{DateTo}{Optionally set date to stop acquiring data (in the format 'YYYY/MM/DD')}

\item{GameSegment}{Optionally filter by 'First+Half', Second+Half', or 'Overtime'.}

\item{LastNGames}{Optionally filter by the last 'n' number of games.}

\item{LeagueID}{Character vector with any combination of '00', '10', and/or '20'.
These codes stand for NBA, WNBA and D-League, respectively.}

\item{Location}{Optionally filter by 'Home' or 'Road' games?}

\item{Month}{Optionally filter by Month (use '1' for Oct, '2' for Nov, etc.)}

\item{OpponentTeamID}{Optionally filter by opposing team}

\item{Outcome}{Optionally filter by wins (use 'W')  or losses (use 'L')}

\item{Period}{Optionally filter by quarter/OT (use '1' for 1st Quarter, '2' for 2nd, etc.)}

\item{Season}{Required filter on season year}

\item{SeasonSegment}{Optionally filter by 'Pre All-Star' or 'Post All-Star'}

\item{SeasonType}{Required filter on either 'Regular Season' or 'Playoffs'}

\item{TeamID}{}

\item{VsConference}{}

\item{VsDivision}{}

\item{PerGame}{Aggregate 'PerGame' or 'Totals'. Only relevant for "dashboards"}
}
\description{
All the arguments to this function
}
\examples{
\dontrun{
# Defaults to Lebron's shots from the 2013-14 regular season
shots <- acquire()
# Lebron's rebounds from the 2013-14 regular season
rebounds <- acquire("rebound")
# The rest of these data types are "dashboards"
# That is, they show summary statistics that will change over
# the season. For this reason, you probably don't want to
# store these in a database.
pass_dash <- acquire("pass")
defense_dash <- acquire("shotdefend")
reb_dash <- acquire("reb")
shot_dash <- acquire("shots")

# All the shots for the 2013-14 season!
ids <- players[, 'PlayerID']
lshots <- lapply(ids, function(x) acquire(PlayerID = x))
allshots <- do.call("rbind", lshots)

# All the rebounds for the 2013-14 season!
lrebounds <- lapply(ids, function(x)
                 acquire(type = "rebound", PlayerID = x))
allrebs <- do.call("rbind", lrebounds)
}
}

