% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fields.R
\docType{data}
\name{fields}
\alias{fields}
\title{Master list of field names and types used to define table formats}
\description{
Master list of field names and types used to define table formats
}
\author{
Carson Sievert
}

