% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{players}
\alias{players}
\title{All available player names and IDs}
\description{
The IDs can be passed along to the PlayerID argument to the
acquire function. To see how this was created inspect the
\code{system.file("scripts/getPlayerIDs.R", package="bbscrapeR")}
}
\author{
Carson Sievert
}

