% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/players.R
\docType{data}
\name{players}
\alias{players}
\title{All available player names and IDs}
\description{
The IDs can be passed along to the PlayerID argument to the
acquire function. To see how this was created inspect the
\code{system.file("scripts/getPlayerIDs.R", package="bbscrapeR")}
}
\author{
Carson Sievert
}

